requirements.txt
openai
chroma
langchain
ffmpeg-python
pydub
whisper
dotenv
chromadb
assemblyai
